+++ YACReaderLibrary/db_helper.h
@@ -37,7 +37,7 @@ public:
     static QList<ComicDB> getFavorites(qulonglong libraryId);
     static QList<ComicDB> getReading(qulonglong libraryId);
     static QList<ReadingList> getReadingLists(qulonglong libraryId);
-    static QList<ComicDB> getReadingListFullContent(qulonglong libraryId, qulonglong readingListId);
+    static QList<ComicDB> getReadingListFullContent(qulonglong libraryId, qulonglong readingListId, bool getFullComicInfoFields = false);
 
     //objects management
     //deletes
@@ -96,6 +96,7 @@ public:
     static ComicDB loadComic(qulonglong id, QSqlDatabase &db);
     static ComicDB loadComic(QString cname, QString cpath, QString chash, QSqlDatabase &database);
     static ComicInfo loadComicInfo(QString hash, QSqlDatabase &db);
+    static ComicInfo getComicInfoFromQuery(QSqlQuery &query, const QString &idKey = "id");
     static QList<QString> loadSubfoldersNames(qulonglong folderId, QSqlDatabase &db);
     //queries
     static bool isFavoriteComic(qulonglong id, QSqlDatabase &db);