+++ YACReaderLibrary/library_window.h
@@ -218,6 +218,8 @@ public:
 
     QAction *showEditShortcutsAction;
 
+    QAction *quitAction;
+
     QAction *updateFolderAction;
     QAction *updateCurrentFolderAction;
 
@@ -313,7 +315,6 @@ public slots:
     void selectSubfolder(const QModelIndex &mi, int child);
     void checkEmptyFolder();
     void openComic();
-    void openComic(const ComicDB &comic);
     void createLibrary();
     void create(QString source, QString dest, QString name);
     void showAddLibrary();