+++ common/yacreader_global.h
@@ -2,8 +2,9 @@
 #define __YACREADER_GLOBAL_H
 
 #include <QStandardPaths>
+#include <QDataStream>
 
-#define VERSION "9.8.0"
+#define VERSION "9.8.1"
 
 #define REMOTE_BROWSE_PERFORMANCE_WORKAROUND "REMOTE_BROWSE_PERFORMANCE_WORKAROUND"
 
@@ -54,9 +55,26 @@ enum LabelColors {
     YDark
 };
 
+struct OpenComicSource {
+    enum Source {
+        Folder = 0,
+        ReadingList
+    };
+
+    Source source;
+    qulonglong sourceId;
+};
+
+QDataStream &operator<<(QDataStream &stream, const OpenComicSource &source);
+QDataStream &operator>>(QDataStream &stream, OpenComicSource &source);
+
 QString getSettingsPath();
 QString colorToName(LabelColors colors);
 QString labelColorToRGBString(LabelColors color);
 
 }
+
+Q_DECLARE_METATYPE(YACReader::OpenComicSource::Source)
+Q_DECLARE_METATYPE(YACReader::OpenComicSource)
+
 #endif